Integrated building state machines from type
